ABC389のC問題のコードを変更した(xの型をintからlonglongに)

+++ ABC/ABC389_C.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main() {
 	int q;
 	cin >> q;
-	vector<int>x(q + 5);
+	vector<long long>x(q + 5);
 	// front = 先頭の位置, back = 末尾の位置
 	int front = 0, back = 0;
 	// 今まで追加してきた累積和の値
